Abstract api
